Added qkeras confusion matrix code.
Added a script just_qkeras_predict.py that takes in the results
from ROUTINE_DNN and dynamic_parseDNNresults and then runs the
yprofiles through the qkeras model from the filter github,
and generates a confusion matrix (based on qkeras).

Added a bash script that automatically runs the dnn analysis
with an input data path.

Comments at the beginning of the scripts indicate usage.